# api/openapi.yaml
openapi: 3.1.0
info:
  title: Asktrix CRM — Mobile Agent API
  version: 0.1.0
  description: |
    Contract between the Asktrix CRM backend and the Android agent app.

    **STATUS: PROPOSAL.** The Asktrix CRM exposes no APIs as of 2026-07-26. This document is written
    as the contract for the CRM team to implement. The mobile app is built against it and against a
    mock server. Nothing here should be read as describing an endpoint that exists today.

    ## Non-negotiable privacy invariant

    **No schema in this document contains a full customer phone number or email address.** Contact
    details exist only as pre-masked strings (`phoneMasked`, `emailMasked`) produced server-side.

    This is deliberate and load-bearing (§4, ADR-0003). The threat model is an insider with a company
    device in hand all day. Client-side masking is defeated in minutes with an HTTP proxy, so the
    unmasked value must never leave the server. Calls are placed by `clientId`; the CRM resolves the
    real number and hands it to the telephony provider, never to the device.

    Reviewers: adding a `phone` or `email` field to any response schema breaks §4. Do not do it.

    ## Conventions

    - All timestamps are RFC 3339 UTC. Clients must not trust the device clock for ordering; the
      server is authoritative.
    - Every mutating request carries an `Idempotency-Key` header. Replaying a key returns the
      original result rather than performing the action twice (§9, §23).
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate an employee and bind the session to this device
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: |
            Credentials are valid but this device is not permitted — unenrolled device, integrity
            failure, or the employee is already bound to a different device.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for a new token pair
      description: |
        Refresh tokens are single-use and rotate on every call. Presenting a previously-used refresh
        token indicates theft: the server must revoke the entire token family and return 401.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken, deviceId]
              properties:
                refreshToken:
                  type: string
                deviceId:
                  type: string
      responses:
        '200':
          description: New token pair issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/session:
    get:
      tags: [Auth]
      summary: Current employee, role and permissions
      responses:
        '200':
          description: Session detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ------------------------------------------------------------- Clients ----
  /clients:
    get:
      tags: [Clients]
      summary: Clients assigned to the authenticated employee (§12)
      description: |
        Returns **only** clients assigned to the caller. Server-side authorisation, never a client
        filter — the device must never receive a record it is not entitled to see.
      parameters:
        - name: status
          in: query
          description: Filter by process status
          schema:
            $ref: '#/components/schemas/ProcessStatus'
        - name: needsFollowUp
          in: query
          description: Only clients with a follow-up due on or before today
          schema:
            type: boolean
        - name: query
          in: query
          description: Free-text search over client name and client ID only — never contact details
          schema:
            type: string
            maxLength: 64
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Assigned clients
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientSummaryPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /clients/{clientId}:
    get:
      tags: [Clients]
      summary: Full client detail with masked contact fields (§3, §4)
      parameters:
        - $ref: '#/components/parameters/ClientId'
      responses:
        '200':
          description: Client detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /clients/{clientId}/status:
    post:
      tags: [Clients]
      summary: Apply a quick status update (§13)
      description: |
        Backs the quick-action buttons: Documents Received, Client Not Responding, Payment Received,
        Waiting Government Approval, Completed, Callback Scheduled.
      parameters:
        - $ref: '#/components/parameters/ClientId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusUpdateRequest'
      responses:
        '200':
          description: Status applied; the resulting timeline entry is returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: |
            The client was modified since `expectedVersion`. The response carries the current server
            state so the client can resolve the conflict without a second round trip.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictResponse'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /clients/{clientId}/remarks:
    post:
      tags: [Clients]
      summary: Add an internal remark
      parameters:
        - $ref: '#/components/parameters/ClientId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body:
                  type: string
                  minLength: 1
                  maxLength: 2000
                recordedAt:
                  type: string
                  format: date-time
                  description: Device time the remark was written; the server records its own receipt time
      responses:
        '201':
          description: Remark recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineEntry'
        '422':
          $ref: '#/components/responses/ValidationFailed'

  # ------------------------------------------------------------ Timeline ----
  /clients/{clientId}/timeline:
    get:
      tags: [Timeline]
      summary: Interaction timeline — calls, remarks, status changes, payments, follow-ups (§8)
      parameters:
        - $ref: '#/components/parameters/ClientId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Timeline entries, newest first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelinePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # --------------------------------------------------------------- Calls ----
  /calls:
    post:
      tags: [Calls]
      summary: Request a click-to-call (§5)
      description: |
        The device sends only a `clientId`. The CRM resolves the customer number server-side and asks
        the telephony provider to bridge the two legs (ADR-0002).

        **The response must never contain the customer's number.** The device does not dial and has
        no telephony permissions.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientId]
              properties:
                clientId:
                  type: string
                serviceId:
                  type: string
                  description: Optional service context, when the client has more than one
                reason:
                  type: string
                  maxLength: 120
                  description: Non-sensitive call purpose, recorded on the timeline
      responses:
        '202':
          description: |
            Call request accepted. The provider places the agent leg first; the outcome arrives by
            FCM push, or by polling this session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: This employee already has a call in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: The telephony provider rejected or could not place the call
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /calls/{callSessionId}:
    get:
      tags: [Calls]
      summary: Current state of a call session
      parameters:
        - name: callSessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Call session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallSession'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /device/push-token:
    put:
      tags: [Device]
      summary: Register or update the FCM registration token (§24)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '204':
          description: Token stored
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from the previous page's `nextCursor`
      schema:
        type: string
...
  schemas:
...
    LoginRequest:
      type: object
      required: [employeeCode, password, device]
      properties:
        employeeCode:
          type: string
          minLength: 1
          maxLength: 64
        password:
          type: string
          format: password
          minLength: 8
        device:
          $ref: '#/components/schemas/DeviceBinding'
...